fix(planner): drop openai from lockfile and move chat types to /types
The previous commit removed the openai dependency from package.json (and
replaced the SDK with a direct fetch) but did not regenerate yarn.lock,
which caused 'yarn install --immutable' to reject the inconsistent state
in CI. This commit:

- Regenerates yarn.lock so it no longer references openai.
- Moves the ChatMessage / ChatCompletionRequest / ChatCompletionResponse
  interfaces out of lib/planner/client.ts and into types/planner.ts to
  satisfy the project's 'no-restricted-syntax' rule (all interfaces in
  /types).

No behavior change.
