+++ types/planner.ts
@@ -124,3 +124,53 @@ export type PlannerFormStatus =
       toolCalls: number;
       modelId: string;
     };
+
+// --- GitHub Models / OpenAI-compatible chat completion shapes ---
+
+export interface ChatToolCall {
+  id: string;
+  type: "function";
+  function: { name: string; arguments: string };
+}
+
+export type ChatMessage =
+  | { role: "system"; content: string }
+  | { role: "user"; content: string }
+  | {
+      role: "assistant";
+      content: string | null;
+      tool_calls?: ChatToolCall[];
+    }
+  | { role: "tool"; tool_call_id: string; content: string };
+
+export interface ChatCompletionToolDefinition {
+  type: "function";
+  function: {
+    name: string;
+    description: string;
+    parameters: Record<string, unknown>;
+  };
+}
+
+export interface ChatCompletionRequest {
+  model: string;
+  messages: ChatMessage[];
+  tools?: ChatCompletionToolDefinition[];
+  temperature?: number;
+  max_tokens?: number;
+}
+
+export interface ChatCompletionChoice {
+  index: number;
+  message: {
+    role: "assistant";
+    content: string | null;
+    tool_calls?: ChatToolCall[];
+  };
+  finish_reason: string | null;
+}
+
+export interface ChatCompletionResponse {
+  id: string;
+  choices: ChatCompletionChoice[];
+}
